requirements.txt
customtkinter
CTkColorPicker
bleak
pillow